requirements.txt
pillow
streamdeck